Embellish exceptions that leave message handlers more details
NServiceBus already logs completely failed messages by default, it is unnecessary to replicate this behaviour.  NServiceBus also only logs the final failure as Error, all retries are either Info or Warnings.

Instead, include the message name and origination endpoint that would have been logged in the exception that bubbles up to NServiceBus so it is still visible for easy diagnostics.